commit-2-8-2013-2337
worked on the 3 object type logins (admin, bus_personnel and commuter).
sorted out a bt of the public interface. have to finish up issues with
the admin+bus_personnel area. also, must work on shortest path algo for
the bus stops and integrating and displaying the route data for bus
routes.